# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conflict Economic Impact Simulator",
    "description": "Local-projection estimates of how armed conflict is associated with GDP per capita, FDI and poverty, plus a Monte Carlo scenario simulator. Estimates are associational, not causal.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/findings": {
      "get": {
        "summary": "Findings",
        "description": "Generated findings for the scrolling page (python -m conflict_sim.findings).",
        "operationId": "findings_findings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Findings_"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/map": {
      "get": {
        "summary": "Map Layer",
        "operationId": "map_layer_map_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "intensity_12m"
              ],
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__MapValue__"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/country/{iso3}/summary": {
      "get": {
        "summary": "Country Summary",
        "operationId": "country_summary_country__iso3__summary_get",
        "parameters": [
          {
            "name": "iso3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iso3"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CountrySummary_"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacktestSummary": {
        "properties": {
          "n_episodes": {
            "type": "integer",
            "title": "N Episodes",
            "description": "past wars replayed"
          },
          "hit_rate_90": {
            "type": "number",
            "title": "Hit Rate 90",
            "description": "share whose actual GDP fell inside the 90% range"
          },
          "median_abs_log_err": {
            "type": "number",
            "title": "Median Abs Log Err"
          },
          "median_abs_log_err_no_war_model": {
            "type": "number",
            "title": "Median Abs Log Err No War Model",
            "description": "same error for a model that assumes the war had no effect"
          }
        },
        "type": "object",
        "required": [
          "n_episodes",
          "hit_rate_90",
          "median_abs_log_err",
          "median_abs_log_err_no_war_model"
        ],
        "title": "BacktestSummary"
      },
...
      "CountrySummary": {
        "properties": {
          "iso3": {
            "type": "string",
            "title": "Iso3"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "population": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "status": {
            "type": "string",
            "enum": [
              "at_war",
              "low_level",
              "calm",
              "unknown"
            ],
            "title": "Status",
            "description": "at_war: >= 25 UCDP-equivalent battle deaths in the last 12 months"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "ok",
              "low",
              "none"
            ],
            "title": "Confidence"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetricRow"
            },
            "type": "object",
            "title": "Metrics"
          },
          "fatalities_36m": {
            "items": {
              "$ref": "#/components/schemas/FatalityPoint"
            },
            "type": "array",
            "title": "Fatalities 36M"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "iso3",
          "name",
          "population",
          "status",
          "as_of",
          "confidence",
          "headline",
          "metrics",
          "fatalities_36m",
          "sources"
        ],
        "title": "CountrySummary"
      },
...
      "Envelope_CountrySummary_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CountrySummary"
          },
          "coverage": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IndicatorCoverage"
            },
            "type": "object",
            "title": "Coverage",
            "description": "keyed by WB indicator code"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warn"
            },
            "type": "array",
            "title": "Warnings"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "coverage",
          "warnings",
          "meta"
        ],
        "title": "Envelope[CountrySummary]"
      },
...
      "Envelope_Findings_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Findings"
          },
          "coverage": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IndicatorCoverage"
            },
            "type": "object",
            "title": "Coverage",
            "description": "keyed by WB indicator code"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warn"
            },
            "type": "array",
            "title": "Warnings"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "coverage",
          "warnings",
          "meta"
        ],
        "title": "Envelope[Findings]"
      },
...
      "Envelope_dict_str__MapValue__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MapValue"
            },
            "type": "object",
            "title": "Data"
          },
          "coverage": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IndicatorCoverage"
            },
            "type": "object",
            "title": "Coverage",
            "description": "keyed by WB indicator code"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warn"
            },
            "type": "array",
            "title": "Warnings"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "coverage",
          "warnings",
          "meta"
        ],
        "title": "Envelope[dict[str, MapValue]]"
      },
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorBody"
      },
      "FatalityPoint": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month",
            "examples": [
              "2026-08"
            ]
          },
          "fatalities": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fatalities",
            "description": "null before ACLED covers the country"
          }
        },
        "type": "object",
        "required": [
          "month",
          "fatalities"
        ],
        "title": "FatalityPoint"
      },
      "Findings": {
        "properties": {
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "estimated_on": {
            "type": "string",
            "title": "Estimated On"
          },
          "acled_as_of": {
            "type": "string",
            "title": "Acled As Of"
          },
          "lowo_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowo On"
          },
          "hero": {
            "items": {
              "$ref": "#/components/schemas/HeroNumber"
            },
            "type": "array",
            "title": "Hero"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/FindingsSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "estimated_on",
          "acled_as_of",
          "lowo_on",
          "hero",
          "sections",
          "sources"
        ],
        "title": "Findings"
      },
      "FindingsSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "heading": {
            "type": "string",
            "title": "Heading",
            "description": "sentence case"
          },
          "sentences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sentences",
            "description": "generated from templates, never hand-written"
          },
          "chart": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart",
            "description": "chart payload; 'type' selects the chart"
          },
          "how_we_know": {
            "type": "string",
            "title": "How We Know",
            "description": "N, caveats and estimation date"
          },
          "sr_summary": {
            "type": "string",
            "title": "Sr Summary",
            "description": "text alternative for the chart"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "heading",
          "sentences",
          "chart",
          "how_we_know",
          "sr_summary"
        ],
        "title": "FindingsSection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeroNumber": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "display": {
            "type": "string",
            "title": "Display",
            "description": "rounded for display, e.g. '120,000'"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          }
        },
        "type": "object",
        "required": [
          "id",
          "value",
          "display",
          "text",
          "as_of"
        ],
        "title": "HeroNumber"
      },
...
      "IndicatorCoverage": {
        "properties": {
          "n_obs": {
            "type": "integer",
            "title": "N Obs",
            "description": "non-null observations (country-years)"
          },
          "first_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Year"
          },
          "last_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Year"
          },
          "share": {
            "type": "number",
            "title": "Share",
            "description": "share of sample years (1990-2023) with data, 0-1"
          }
        },
        "type": "object",
        "required": [
          "n_obs",
          "first_year",
          "last_year",
          "share"
        ],
        "title": "IndicatorCoverage"
      },
      "MapValue": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "lo": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lo",
            "description": "5th percentile (equals value for observed metrics)"
          },
          "hi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hi",
            "description": "95th percentile"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "ok",
              "low",
              "none"
            ],
            "title": "Confidence"
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "description": "ACLED data date, YYYY-MM-DD"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "why a value is null, if it is"
          }
        },
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi",
          "confidence",
          "as_of"
        ],
        "title": "MapValue"
      },
      "Meta": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "ingestion_pipeline | worldbank_dev_pull | synthetic_fixture"
          },
          "artifacts_created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Created At"
          },
          "n_draws": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Draws"
          },
          "backtest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Associational estimates (two-way FE local projections), not causal effects."
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "Meta"
      },
...
      "MetricRow": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "lo": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lo",
            "description": "5th percentile (equals value for observed metrics)"
          },
          "hi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hi",
            "description": "95th percentile"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "ok",
              "low",
              "none"
            ],
            "title": "Confidence"
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "description": "ACLED data date, YYYY-MM-DD"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "why a value is null, if it is"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs",
            "description": "data that went into this number"
          }
        },
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi",
          "confidence",
          "as_of",
          "inputs"
        ],
        "title": "MetricRow"
      },
...
      "Source": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "use": {
            "type": "string",
            "title": "Use"
          }
        },
        "type": "object",
        "required": [
          "name",
          "detail",
          "use"
        ],
        "title": "Source"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Warn": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning"
            ],
            "title": "Severity",
            "default": "warning"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Warn"
      }
    }
  }
}